+++ interface/lib/definitions.dart
@@ -1,4 +1,4 @@
-/// Uses for defining standard of data.
+/// Uses for defining standard of data and collections.
 library definitions;
 
 export 'src/definitions/bios_dateformat.dart';