+++ interface/CHANGELOG.md
@@ -1,5 +1,7 @@
-## 3.1.0
+## 4.0.0
 
+* Migrate to `Map` based dictionary implementations.
+* Drop entire stringify dictionary function.
 * Define new exceptions for creating unsupported platform in `DeviceVendorInfoDictionary`.
 
 ## 3.0.0